requirements.txt
pandas
matplotlib
streamlit
pywebview